feat: move to sync (none ETL) implementation - remove ruffus
